requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
atlassian-python-api==3.39.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses-json==0.5.8
Deprecated==1.2.14
frozenlist==1.3.3
fsspec==2023.6.0
idna==3.4
langchain==0.0.188
langchainplus-sdk==0.0.10
llama-index==0.6.27
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
pydantic==1.10.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
SQLAlchemy==2.0.16
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
yarl==1.9.2
